+++ packages/core/src/index.ts
@@ -18,5 +18,6 @@ export * from "./invoice";
 export * from "./money";
 export * from "./orders";
 export * from "./profitability";
+export * from "./savedPlaces";
 export * from "./schemas";
 export * from "./vehicleCosts";